+++ src/RollbarDotNet/Builder/ConfigurationBuilder.cs
@@ -1,31 +1,32 @@
 ﻿namespace RollbarDotNet.Builder
 {
-    using Microsoft.Extensions.Configuration;
+    using Configuration;
+    using Microsoft.Extensions.Options;
     using Payloads;
     using System;
 
     public class ConfigurationBuilder : IBuilder
     {
-        public ConfigurationBuilder(IConfigurationRoot configuration)
+        public ConfigurationBuilder(IOptions<RollbarOptions> configuration)
         {
-            this.Configuration = configuration;
+            this.Configuration = configuration.Value;
         }
 
-        protected IConfigurationRoot Configuration { get; set; }
+        protected RollbarOptions Configuration { get; set; }
 
         public void Execute(Payload payload)
         {
-            payload.AccessToken = this.Configuration["Rollbar:AccessToken"];
-            payload.Data.Environment = this.Configuration["Rollbar:Environment"];
+            payload.AccessToken = this.Configuration.AccessToken;
+            payload.Data.Environment = this.Configuration.Environment;
 
             if(string.IsNullOrEmpty(payload.AccessToken))
             {
-                throw new InvalidOperationException("Configuration variable Rollbar:AccessToken must be set.");
+                throw new InvalidOperationException("Configuration variable for your Rollbar AccessToken must be set (did you include services.Configure<RollbarOptions>?).");
             }
 
             if (string.IsNullOrEmpty(payload.Data.Environment))
             {
-                throw new InvalidOperationException("Configuration variable Rollbar:Environment must be set.");
+                throw new InvalidOperationException("Configuration variable for your Rollbar Environment must be set (did you include services.Configure<RollbarOptions>?).");
             }
         }
     }