+++ src/RollbarDotNet/Configuration/RollbarOptions.cs
@@ -1,0 +1,11 @@
+﻿namespace RollbarDotNet.Configuration
+{
+    public class RollbarOptions
+    {
+        public string AccessToken { get; set; }
+
+        public string Environment { get; set; }
+
+        public BlacklistConfiguration Blacklist { get; set; }
+    }
+}