+++ src/RollbarDotNet/Core/ServiceExtensions.cs
@@ -4,12 +4,15 @@
     using Blacklisters;
     using Builder;
     using Configuration;
+    using Microsoft.AspNetCore.Http;
     using Microsoft.Extensions.Configuration;
     using Microsoft.Extensions.DependencyInjection;
+    using Microsoft.Extensions.DependencyInjection.Extensions;
+    using Microsoft.Extensions.Options;
 
     public static class ServiceExtensions
     {
-        public static IServiceCollection AddRollbar(this IServiceCollection services, IConfigurationRoot configuration)
+        public static IServiceCollection AddRollbar(this IServiceCollection services)
         {
             return services
                 .AddSingleton<IBuilder, Builder.ConfigurationBuilder>()
@@ -19,16 +22,17 @@
                 .AddSingleton<IEnvironment, SystemEnvironment>()
                 .AddSingleton<IBlacklister, ConfigurationBlacklister>()
                 .AddSingleton<IBlacklistCollection, BlacklistCollection>()
-                .Configure<BlacklistConfiguration>(configuration.GetSection("Rollbar:Blacklist"))
                 .AddScoped<Rollbar>();
         }
 
-        public static IServiceCollection AddRollbarWeb(this IServiceCollection services, IConfigurationRoot configuration)
+        public static IServiceCollection AddRollbarWeb(this IServiceCollection services)
         {
-            return AddRollbar(services, configuration)
+            services.TryAddSingleton<IHttpContextAccessor, HttpContextAccessor>();
+            return services.AddRollbar()
                 .AddSingleton<IBuilder, ServerBuilder>()
                 .AddScoped<IBuilder, RequestBuilder>()
                 .AddScoped<IBuilder, PersonBuilder>();
+
         }
     }
 }